Add ability to retest
Adds ability to re-run tests if required.

Ignores if only markdown files are changed as they are handled by tox and weren't touched by these tests.

Adds basic Refurb settings to pyproject. Not adding as dev requirement as it's a nice to run but by no means needed.

Updates black and isort versions.